backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pandas==2.2.2
scikit-learn==1.5.0
joblib==1.4.2
python-multipart==0.0.9
numpy==1.26.4
httpx==0.27.0
python-dotenv==1.0.1  # Add this line